EKF: Use ecl:powf for integer exponents in EKFGSF_Yaw
Also improve protection against badly conditioned calculation
