+++ EKF/common.h
@@ -43,6 +43,28 @@
 
 #include <matrix/math.hpp>
 
+namespace ecl{
+	inline float powf(float x, int exp)
+	{
+		float ret;
+		if (exp > 0) {
+			ret = x;
+			for (int count = 1; count < exp; count++) {
+				ret *= x;
+			}
+			return ret;
+		} else if (exp < 0) {
+			ret = x;
+			for (int count = -1; count > exp; count--) {
+				ret *= x;
+			}
+			return 1.0f / ret;
+		} else {
+			return 1.0f;
+		}
+	}
+}
+
 namespace estimator
 {
