+++ EKF/ekf.h
@@ -44,28 +44,6 @@
 
 #include "estimator_interface.h"
 
-namespace ecl{
-	inline float powf(float x, int exp)
-	{
-		float ret;
-		if (exp > 0) {
-			ret = x;
-			for (int count = 1; count < exp; count++) {
-				ret *= x;
-			}
-			return ret;
-		} else if (exp < 0) {
-			ret = x;
-			for (int count = -1; count > exp; count--) {
-				ret *= x;
-			}
-			return 1.0f / ret;
-		} else {
-			return 1.0f;
-		}
-	}
-}
-
 class Ekf : public EstimatorInterface
 {
 public:
